requirements.txt
opencv-python
numpy
matplotlib
tqdm
pytesseract
pdf2image
scikit-image
pymupdf
python-telegram-bot